Initial upload live script for speech extraction
loads and plots audio signals, cuts to when people start talking, plots the stft spectrum, after melbank freq wrapping, and obtains MLCC. Code optimizations can still be performed as it repeats the loading file and process audio signal part for each stage. OK to start doing the next step though, as the MLCC coefficients are generated, but may need to be optimized further.